refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/folder_tree/electron-folder-visibility.spec.ts into folder
