+++ packages/libraries/graph-tools/src/commands/folders/folderMaterialize.ts
@@ -15,6 +15,8 @@ import { folderMaterializeImplementation } from './folderMaterialize/implementat
 
 const {
   DEFAULT_TIMEOUT_MS,
+  createScratchFixture,
+  folderMaterialize,
   waitForGraphReady,
   withTrailingSlash,
 } = folderMaterializeImplementation
@@ -75,16 +77,7 @@ export type FolderMaterializeResult = {
   writeFolder: string
 }
 
-export async function createScratchFixture(writeFolder: string): Promise<ScratchFixture> {
-  return folderMaterializeImplementation.createScratchFixture(writeFolder)
-}
-
-export async function folderMaterialize(
-  page: Parameters<typeof folderMaterializeImplementation.folderMaterialize>[0],
-  opts: Parameters<typeof folderMaterializeImplementation.folderMaterialize>[1],
-): Promise<Response<FolderMaterializeResult>> {
-  return folderMaterializeImplementation.folderMaterialize(page, opts)
-}
+export { createScratchFixture, folderMaterialize }
 
 function usage(message?: string): Response<never> {
   return err(
